+++ common/src/main/java/com/illusivesoulworks/charmofundying/common/TotemProviders.java
@@ -21,10 +21,10 @@ package com.illusivesoulworks.charmofundying.common;
 import com.google.common.collect.ImmutableSet;
 import com.illusivesoulworks.charmofundying.common.integration.BMEnchantedTotemEffectProvider;
 import com.illusivesoulworks.charmofundying.platform.Services;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Predicate;
 import net.minecraft.world.item.Item;
 
@@ -37,7 +37,8 @@ public class TotemProviders {
     }
   };
 
-  private static final Map<String, ITotemEffectProvider> EFFECT_PROVIDERS = new HashMap<>();
+  private static final Map<String, ITotemEffectProvider> EFFECT_PROVIDERS =
+      new ConcurrentHashMap<>();
 
   public static void init() {
     EFFECT_PROVIDERS.put("minecraft:totem_of_undying", new VanillaTotemEffectProvider());
@@ -53,5 +54,9 @@ public class TotemProviders {
 
   public static Optional<ITotemEffectProvider> getEffectProvider(final Item item) {
     return Optional.ofNullable(EFFECT_PROVIDERS.get(Services.PLATFORM.getRegistryName(item)));
+  }
+
+  public static void putEffectProvider(String key, ITotemEffectProvider provider) {
+    EFFECT_PROVIDERS.put(key, provider);
   }
 }