+++ fabric/src/main/java/com/illusivesoulworks/charmofundying/platform/FabricPlatform.java
@@ -18,13 +18,20 @@
 
 package com.illusivesoulworks.charmofundying.platform;
 
+import com.illusivesoulworks.charmofundying.CharmOfUndyingConstants;
 import com.illusivesoulworks.charmofundying.common.TotemProviders;
+import com.illusivesoulworks.charmofundying.common.network.SPacketUseTotem;
 import com.illusivesoulworks.charmofundying.platform.services.IPlatform;
 import dev.emi.trinkets.api.SlotReference;
 import dev.emi.trinkets.api.TrinketsApi;
 import java.util.List;
+import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
+import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.core.Registry;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.Tuple;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
@@ -49,5 +56,19 @@ public class FabricPlatform implements IPlatform {
   @Override
   public boolean isModLoaded(String name) {
     return FabricLoader.getInstance().isModLoaded(name);
+  }
+
+  @Override
+  public void broadcastTotemEvent(LivingEntity livingEntity) {
+    FriendlyByteBuf buf = PacketByteBufs.create();
+    SPacketUseTotem.encode(new SPacketUseTotem(livingEntity.getId()), buf);
+
+    for (ServerPlayer serverPlayer : PlayerLookup.tracking(livingEntity)) {
+      ServerPlayNetworking.send(serverPlayer, CharmOfUndyingConstants.TOTEM_EVENT, buf);
+    }
+
+    if (livingEntity instanceof ServerPlayer serverPlayer) {
+      ServerPlayNetworking.send(serverPlayer, CharmOfUndyingConstants.TOTEM_EVENT, buf);
+    }
   }
 }