+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 Prior to version 6.0.0, this project used MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [6.1.0+1.19.2] - 2023.01.14
+### Added
+- Added integration with Fabric Waystone's Void Totem [#13](https://github.com/illusivesoulworks/charmofundying/issues/13)
+### Fixed
+- Fixed totem animation using vanilla totem instead of modded totems when applicable
+
 ## [6.0.0+1.19.2] - 2022.08.09
 ### Changed
 - Updated to Minecraft 1.19.2